# otelgrpc: reduce allocation for stats handler