Good day,

While I was looking at some profiles I noticed that handleRPC was doing a bit more allocations then expected.
This PR avoids allocating metricAttrs for GRPC `stats.Begin`, `stats.OutHeader`, `stats.OutTrailer` and the default case.

Thanks for reviewing and let's me know what you think :smile:  

Benchstats of https://github.com/open-telemetry/opentelemetry-go-contrib/pull/7186/commits/02d373d407f09fe088b94ac0ef1a26a75a08d264 + https://github.com/open-telemetry/opentelemetry-go-contrib/pull/7186/commits/93abfa2563a2d080ac5c2ef8ef29a80a04fc71f0: 
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc
cpu: AMD Ryzen 7 Pro 7735U with Radeon Graphics     
                                           │   base.txt   │                new.txt                │
                                           │    sec/op    │   sec/op     vs base                  │
ServerHandler_HandleRPC_Begin-16              87.08n ± 1%   18.24n ± 2%  -79.05% (p=0.000 n=10)
ServerHandler_HandleRPC_InPayload-16          676.2n ± 1%   663.3n ± 1%   -1.90% (p=0.000 n=10)
ServerHandler_HandleRPC_OutPayload-16         679.9n ± 1%   671.5n ± 1%        ~ (p=0.079 n=10)
ServerHandler_HandleRPC_OutTrailer-16         88.97n ± 1%   18.45n ± 0%  -79.26% (p=0.000 n=10)
ServerHandler_HandleRPC_OutHeader-16          428.1n ± 3%   343.7n ± 1%  -19.72% (p=0.000 n=10)
ServerHandler_HandleRPC_End-16                410.9n ± 1%   351.1n ± 1%  -14.55% (p=0.000 n=10)
ServerHandler_HandleRPC_Nil-16                88.03n ± 1%   17.97n ± 0%  -79.58% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_Begin-16         88.54n ± 1%   18.41n ± 1%  -79.21% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_InPayload-16     382.1n ± 0%   246.4n ± 2%  -35.51% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutPayload-16    377.0n ± 2%   245.4n ± 2%  -34.89% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutTrailer-16    88.47n ± 0%   18.42n ± 1%  -79.18% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutHeader-16    308.95n ± 1%   19.50n ± 1%  -93.69% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_End-16           369.6n ± 1%   336.8n ± 1%   -8.87% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_Nil-16           88.52n ± 1%   18.01n ± 0%  -79.65% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp-16               1.834µ ± 1%
ServerHandler_HandleRPC-16                    1.917µ ± 1%
geomean                                       287.8n        83.41n       -62.12%                ¹
¹ benchmark set differs from baseline; geomeans may not be comparable

                                           │   base.txt   │                  new.txt                  │
                                           │     B/op     │    B/op     vs base                       │
ServerHandler_HandleRPC_Begin-16               256.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_InPayload-16          1032.0 ± 0%   968.0 ± 0%    -6.20% (p=0.000 n=10)
ServerHandler_HandleRPC_OutPayload-16         1032.0 ± 0%   968.0 ± 0%    -6.20% (p=0.000 n=10)
ServerHandler_HandleRPC_OutTrailer-16          256.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_OutHeader-16           521.0 ± 0%   265.0 ± 0%   -49.14% (p=0.000 n=10)
ServerHandler_HandleRPC_End-16                 608.0 ± 0%   544.0 ± 0%   -10.53% (p=0.000 n=10)
ServerHandler_HandleRPC_Nil-16                 256.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_Begin-16          256.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_InPayload-16      776.0 ± 0%   416.0 ± 0%   -46.39% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutPayload-16     776.0 ± 0%   416.0 ± 0%   -46.39% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutTrailer-16     256.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutHeader-16      416.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_End-16            608.0 ± 0%   544.0 ± 0%   -10.53% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_Nil-16            256.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp-16              3.266Ki ± 0%
ServerHandler_HandleRPC-16                   3.266Ki ± 0%
geomean                                        579.5                    ?                       ¹ ² ³
¹ benchmark set differs from baseline; geomeans may not be comparable
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                                           │  base.txt  │                  new.txt                  │
                                           │ allocs/op  │ allocs/op   vs base                       │
ServerHandler_HandleRPC_Begin-16             1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_InPayload-16         8.000 ± 0%   8.000 ± 0%         ~ (p=1.000 n=10) ¹
ServerHandler_HandleRPC_OutPayload-16        8.000 ± 0%   8.000 ± 0%         ~ (p=1.000 n=10) ¹
ServerHandler_HandleRPC_OutTrailer-16        1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_OutHeader-16         5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ServerHandler_HandleRPC_End-16               5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ServerHandler_HandleRPC_Nil-16               1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_Begin-16        1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_InPayload-16    7.000 ± 0%   4.000 ± 0%   -42.86% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutPayload-16   7.000 ± 0%   4.000 ± 0%   -42.86% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutTrailer-16   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_OutHeader-16    5.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_End-16          5.000 ± 0%   4.000 ± 0%   -20.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp_Nil-16          1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
ServerHandler_HandleRPC_NoOp-16              27.00 ± 0%
ServerHandler_HandleRPC-16                   27.00 ± 0%
geomean                                      3.734                    ?                       ² ³ ⁴
¹ all samples are equal
² benchmark set differs from baseline; geomeans may not be comparable
³ summaries must be >0 to compute geomean
⁴ ratios must be >0 to compute geomean
```