@dmathieu 2025-04-10
Could you add a benchmark so we can track the number of allocations?