@boekkooi-impossiblecloud 2025-04-10
Hey @dmathieu 

I added a basic benchmark in https://github.com/open-telemetry/opentelemetry-go-contrib/pull/7186/commits/02d373d407f09fe088b94ac0ef1a26a75a08d264.

Let me know what you think :smile: 

Benchstats results are:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc
cpu: AMD Ryzen 7 Pro 7735U with Radeon Graphics     
                                │   base.txt   │               new.txt               │
                                │    sec/op    │   sec/op     vs base                │
ServerHandler_HandleRPC_NoOp-16   1818.0n ± 2%   945.8n ± 2%  -47.98% (p=0.000 n=10)
ServerHandler_HandleRPC-16        1911.0n ± 1%   966.3n ± 3%  -49.43% (p=0.000 n=10)
geomean                            1.864µ        956.0n       -48.71%

                                │   base.txt   │               new.txt                │
                                │     B/op     │     B/op      vs base                │
ServerHandler_HandleRPC_NoOp-16   3.266Ki ± 0%   1.375Ki ± 0%  -57.89% (p=0.000 n=10)
ServerHandler_HandleRPC-16        3.266Ki ± 0%   1.375Ki ± 0%  -57.89% (p=0.000 n=10)
geomean                           3.266Ki        1.375Ki       -57.89%

                                │  base.txt  │              new.txt               │
                                │ allocs/op  │ allocs/op   vs base                │
ServerHandler_HandleRPC_NoOp-16   27.00 ± 0%   14.00 ± 0%  -48.15% (p=0.000 n=10)
ServerHandler_HandleRPC-16        27.00 ± 0%   14.00 ± 0%  -48.15% (p=0.000 n=10)
geomean                           27.00        14.00       -48.15%
```